Update get Articles methode
Create context for get posts
short articles in DOM
